requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-twitter==3.5
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
